@@ -6,26 +6,26 @@ repository: https://github.com/GetStream/stream-chat-flutter
 issue_tracker: https://github.com/GetStream/stream-chat-flutter/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   async: ^2.5.0
   collection: ^1.15.0
-  dio: ">=4.0.0-prev3 <4.0.0"
+  dio: ^4.0.0
   equatable: ^2.0.0
   freezed_annotation: ^0.14.0
   http_parser: ^4.0.0
-  json_annotation: ^4.0.0
-  logging: ^1.0.0
+  json_annotation: ^4.0.1
+  logging: ^1.0.1
   meta: ^1.3.0
   mime: ^1.0.0
   rxdart: ^0.26.0
-  uuid: ^3.0.0
+  uuid: ^3.0.4
   web_socket_channel: ^2.0.0
 
 dev_dependencies:
-  build_runner: ^1.10.0
-  freezed: ^0.14.0
-  json_serializable: ^4.0.0
-  mocktail: ^0.1.0
-  test: ^1.16.0
+  build_runner: ^1.12.2
+  freezed: ^0.14.1+2
+  json_serializable: ^4.1.0
+  mocktail: ^0.1.1
+  test: ^1.16.8
