@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:collection/collection.dart';
 import 'package:logging/logging.dart';
-import 'package:meta/meta.dart';
 import 'package:stream_chat/src/api/channel.dart';
 import 'package:stream_chat/src/api/retry_policy.dart';
 import 'package:stream_chat/src/event_type.dart';
@@ -14,7 +13,7 @@ import 'package:stream_chat/stream_chat.dart';
 class RetryQueue {
   /// Instantiate a new RetryQueue object
   RetryQueue({
-    @required this.channel,
+    required this.channel,
     this.logger,
   }) {
     _retryPolicy = channel.client.retryPolicy;
@@ -28,14 +27,14 @@ class RetryQueue {
   final Channel channel;
 
   /// The logger associated to this queue
-  final Logger logger;
+  final Logger? logger;
 
   final _subscriptions = <StreamSubscription>[];
 
   void _listenConnectionRecovered() {
     _subscriptions
         .add(channel.client.on(EventType.connectionRecovered).listen((event) {
-      if (!_isRetrying && event.online) {
+      if (!_isRetrying && event.online!) {
         _startRetrying();
       }
     }));
@@ -43,12 +42,13 @@ class RetryQueue {
 
   final HeapPriorityQueue<Message> _messageQueue = HeapPriorityQueue(_byDate);
   bool _isRetrying = false;
-  RetryPolicy _retryPolicy;
+  RetryPolicy? _retryPolicy;
 
   /// Add a list of messages
   void add(List<Message> messages) {
     logger?.info('added ${messages.length} messages');
     final messageList = _messageQueue.toList();
+
     _messageQueue.addAll(messages
         .where((element) => !messageList.any((m) => m.id == element.id)));
 
@@ -60,7 +60,7 @@ class RetryQueue {
   Future<void> _startRetrying() async {
     logger?.info('start retrying');
     _isRetrying = true;
-    final retryPolicy = _retryPolicy.copyWith(attempt: 0);
+    final retryPolicy = _retryPolicy!.copyWith(attempt: 0);
 
     while (_messageQueue.isNotEmpty) {
       final message = _messageQueue.first;
@@ -72,7 +72,7 @@ class RetryQueue {
         logger?.info('now ${_messageQueue.length} messages in the queue');
         retryPolicy.attempt = 0;
       } catch (error) {
-        ApiError apiError;
+        ApiError? apiError;
         if (error is DioError) {
           if (error.type == DioErrorType.response) {
             _messageQueue.remove(message);
@@ -84,7 +84,7 @@ class RetryQueue {
           );
         } else if (error is ApiError) {
           apiError = error;
-          if (apiError.status?.toString()?.startsWith('4') == true) {
+          if (apiError.status?.toString().startsWith('4') == true) {
             _messageQueue.remove(message);
             return;
           }
@@ -101,6 +101,7 @@ class RetryQueue {
         }
 
         retryPolicy.attempt++;
+
         final timeout = retryPolicy.retryTimeout(
           channel.client,
           retryPolicy.attempt,
@@ -112,13 +113,13 @@ class RetryQueue {
     _isRetrying = false;
   }
 
-  void _sendFailedEvent(Message message) {
-    final newStatus = message.status == MessageSendingStatus.sending
+  void _sendFailedEvent(Message? message) {
+    final newStatus = message!.status == MessageSendingStatus.sending
         ? MessageSendingStatus.failed
         : (message.status == MessageSendingStatus.updating
             ? MessageSendingStatus.failed_update
             : MessageSendingStatus.failed_delete);
-    channel.state.addMessage(message.copyWith(
+    channel.state!.addMessage(message.copyWith(
       status: newStatus,
     ));
   }
@@ -141,20 +142,24 @@ class RetryQueue {
       final messageList = _messageQueue.toList();
       if (event.message != null) {
         final messageIndex =
-            messageList.indexWhere((m) => m.id == event.message.id);
+            messageList.indexWhere((m) => m.id == event.message!.id);
         if (messageIndex == -1 &&
             [
               MessageSendingStatus.failed_update,
               MessageSendingStatus.failed,
               MessageSendingStatus.failed_delete,
-            ].contains(event.message.status)) {
+            ].contains(event.message!.status)) {
           logger?.info('add message from events');
-          add([event.message]);
+          final m = event.message;
+
+          if (m != null) {
+            add([m]);
+          }
         } else if (messageIndex != -1 &&
             [
               MessageSendingStatus.sent,
               null,
-            ].contains(event.message.status)) {
+            ].contains(event.message!.status)) {
           _messageQueue.remove(messageList[messageIndex]);
         }
       }
@@ -171,10 +176,14 @@ class RetryQueue {
     final date1 = _getMessageDate(m1);
     final date2 = _getMessageDate(m2);
 
+    if (date1 == null || date2 == null) {
+      return 0;
+    }
+
     return date1.compareTo(date2);
   }
 
-  static DateTime _getMessageDate(Message m1) {
+  static DateTime? _getMessageDate(Message m1) {
     switch (m1.status) {
       case MessageSendingStatus.failed_delete:
       case MessageSendingStatus.deleting:
