@@ -12,15 +12,16 @@ class Member {
     this.user,
     this.inviteAcceptedAt,
     this.inviteRejectedAt,
-    this.invited,
+    this.invited = false,
     this.role,
     this.userId,
-    this.isModerator,
-    this.createdAt,
-    this.updatedAt,
-    this.banned,
-    this.shadowBanned,
-  });
+    this.isModerator = false,
+    DateTime? createdAt,
+    DateTime? updatedAt,
+    this.banned = false,
+    this.shadowBanned = false,
+  })  : createdAt = createdAt ?? DateTime.now(),
+        updatedAt = updatedAt ?? DateTime.now();
 
   /// Create a new instance from a json
   factory Member.fromJson(Map<String, dynamic> json) {
@@ -31,30 +32,34 @@ class Member {
   }
 
   /// The interested user
-  final User user;
+  final User? user;
 
   /// The date in which the user accepted the invite to the channel
-  final DateTime inviteAcceptedAt;
+  final DateTime? inviteAcceptedAt;
 
   /// The date in which the user rejected the invite to the channel
-  final DateTime inviteRejectedAt;
+  final DateTime? inviteRejectedAt;
 
   /// True if the user has been invited to the channel
+  @JsonKey(defaultValue: false)
   final bool invited;
 
   /// The role of the user in the channel
-  final String role;
+  final String? role;
 
   /// The id of the interested user
-  final String userId;
+  final String? userId;
 
   /// True if the user is a moderator of the channel
+  @JsonKey(defaultValue: false)
   final bool isModerator;
 
   /// True if the member is banned from the channel
+  @JsonKey(defaultValue: false)
   final bool banned;
 
   /// True if the member is shadow banned from the channel
+  @JsonKey(defaultValue: false)
   final bool shadowBanned;
 
   /// The date of creation
@@ -65,17 +70,17 @@ class Member {
 
   /// Creates a copy of [Member] with specified attributes overridden.
   Member copyWith({
-    User user,
-    DateTime inviteAcceptedAt,
-    DateTime inviteRejectedAt,
-    bool invited,
-    String role,
-    String userId,
-    bool isModerator,
-    DateTime createdAt,
-    DateTime updatedAt,
-    bool banned,
-    bool shadowBanned,
+    User? user,
+    DateTime? inviteAcceptedAt,
+    DateTime? inviteRejectedAt,
+    bool? invited,
+    String? role,
+    String? userId,
+    bool? isModerator,
+    DateTime? createdAt,
+    DateTime? updatedAt,
+    bool? banned,
+    bool? shadowBanned,
   }) =>
       Member(
         user: user ?? this.user,
