@@ -12,19 +12,19 @@ part 'own_user.g.dart';
 class OwnUser extends User {
   /// Constructor used for json serialization
   OwnUser({
-    this.devices,
-    this.mutes,
-    this.totalUnreadCount,
+    this.devices = const [],
+    this.mutes = const [],
+    this.totalUnreadCount = 0,
     this.unreadChannels,
-    this.channelMutes,
-    String id,
-    String role,
-    DateTime createdAt,
-    DateTime updatedAt,
-    DateTime lastActive,
-    bool online,
-    Map<String, dynamic> extraData,
-    bool banned,
+    this.channelMutes = const [],
+    String id = '',
+    String role = '',
+    DateTime? createdAt,
+    DateTime? updatedAt,
+    DateTime? lastActive,
+    bool online = false,
+    Map<String, dynamic> extraData = const {},
+    bool banned = false,
   }) : super(
           id: id,
           role: role,
@@ -41,24 +41,34 @@ class OwnUser extends User {
       Serialization.moveToExtraDataFromRoot(json, topLevelFields));
 
   /// List of user devices
-  @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
+  @JsonKey(
+      includeIfNull: false,
+      toJson: Serialization.readOnly,
+      defaultValue: <Device>[])
   final List<Device> devices;
 
   /// List of users muted by the user
-  @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
+  @JsonKey(
+      includeIfNull: false,
+      toJson: Serialization.readOnly,
+      defaultValue: <Mute>[])
   final List<Mute> mutes;
 
   /// List of users muted by the user
-  @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
+  @JsonKey(
+      includeIfNull: false,
+      toJson: Serialization.readOnly,
+      defaultValue: <Mute>[])
   final List<Mute> channelMutes;
 
   /// Total unread messages by the user
-  @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
+  @JsonKey(
+      includeIfNull: false, toJson: Serialization.readOnly, defaultValue: 0)
   final int totalUnreadCount;
 
   /// Total unread channels by the user
   @JsonKey(includeIfNull: false, toJson: Serialization.readOnly)
-  final int unreadChannels;
+  final int? unreadChannels;
 
   /// Known top level fields.
   /// Useful for [Serialization] methods.
@@ -74,5 +84,6 @@ class OwnUser extends User {
   /// Serialize to json
   @override
   Map<String, dynamic> toJson() => Serialization.moveFromExtraDataToRoot(
-      _$OwnUserToJson(this), topLevelFields);
+        _$OwnUserToJson(this),
+      );
 }
