@@ -1,6 +1,6 @@
 /// Useful extension functions for [Map]
-extension MapX on Map {
+extension MapX<K, V> on Map<K, V> {
   /// Returns a new map with null keys or values removed
-  Map<String, dynamic> get nullProtected =>
-      {...this}..removeWhere((key, value) => key == null || value == null);
+  Map<K, V> get nullProtected =>
+      Map.from(this)..removeWhere((key, value) => key == null || value == null);
 }
