@@ -6,46 +6,37 @@ part of 'attachment.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Attachment _$AttachmentFromJson(Map json) {
+Attachment _$AttachmentFromJson(Map<String, dynamic> json) {
   return Attachment(
-    id: json['id'] as String,
-    type: json['type'] as String,
-    titleLink: json['title_link'] as String,
-    title: json['title'] as String,
-    thumbUrl: json['thumb_url'] as String,
-    text: json['text'] as String,
-    pretext: json['pretext'] as String,
-    ogScrapeUrl: json['og_scrape_url'] as String,
-    imageUrl: json['image_url'] as String,
-    footerIcon: json['footer_icon'] as String,
-    footer: json['footer'] as String,
+    id: json['id'] as String?,
+    type: json['type'] as String?,
+    titleLink: json['title_link'] as String?,
+    title: json['title'] as String?,
+    thumbUrl: json['thumb_url'] as String?,
+    text: json['text'] as String?,
+    pretext: json['pretext'] as String?,
+    ogScrapeUrl: json['og_scrape_url'] as String?,
+    imageUrl: json['image_url'] as String?,
+    footerIcon: json['footer_icon'] as String?,
+    footer: json['footer'] as String?,
     fields: json['fields'],
-    fallback: json['fallback'] as String,
-    color: json['color'] as String,
-    authorName: json['author_name'] as String,
-    authorLink: json['author_link'] as String,
-    authorIcon: json['author_icon'] as String,
-    assetUrl: json['asset_url'] as String,
-    actions: (json['actions'] as List)
-        ?.map((e) => e == null
-            ? null
-            : Action.fromJson((e as Map)?.map(
-                (k, e) => MapEntry(k as String, e),
-              )))
-        ?.toList(),
-    extraData: (json['extra_data'] as Map)?.map(
-      (k, e) => MapEntry(k as String, e),
-    ),
+    fallback: json['fallback'] as String?,
+    color: json['color'] as String?,
+    authorName: json['author_name'] as String?,
+    authorLink: json['author_link'] as String?,
+    authorIcon: json['author_icon'] as String?,
+    assetUrl: json['asset_url'] as String?,
+    actions: (json['actions'] as List<dynamic>?)
+            ?.map((e) => Action.fromJson(e as Map<String, dynamic>))
+            .toList() ??
+        [],
+    extraData: json['extra_data'] as Map<String, dynamic>?,
     file: json['file'] == null
         ? null
-        : AttachmentFile.fromJson((json['file'] as Map)?.map(
-            (k, e) => MapEntry(k as String, e),
-          )),
+        : AttachmentFile.fromJson(json['file'] as Map<String, dynamic>),
     uploadState: json['upload_state'] == null
         ? null
-        : UploadState.fromJson((json['upload_state'] as Map)?.map(
-            (k, e) => MapEntry(k as String, e),
-          )),
+        : UploadState.fromJson(json['upload_state'] as Map<String, dynamic>),
   );
 }
 
@@ -75,10 +66,10 @@ Map<String, dynamic> _$AttachmentToJson(Attachment instance) {
   writeNotNull('author_link', instance.authorLink);
   writeNotNull('author_icon', instance.authorIcon);
   writeNotNull('asset_url', instance.assetUrl);
-  writeNotNull('actions', instance.actions?.map((e) => e?.toJson())?.toList());
+  val['actions'] = instance.actions.map((e) => e.toJson()).toList();
   writeNotNull('file', instance.file?.toJson());
-  writeNotNull('upload_state', instance.uploadState?.toJson());
+  val['upload_state'] = instance.uploadState.toJson();
   writeNotNull('extra_data', instance.extraData);
-  writeNotNull('id', instance.id);
+  val['id'] = instance.id;
   return val;
 }
