@@ -6,7 +6,7 @@ part of 'device.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Device _$DeviceFromJson(Map json) {
+Device _$DeviceFromJson(Map<String, dynamic> json) {
   return Device(
     id: json['id'] as String,
     pushProvider: json['push_provider'] as String,
