@@ -6,64 +6,44 @@ part of 'message.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Message _$MessageFromJson(Map json) {
+Message _$MessageFromJson(Map<String, dynamic> json) {
   return Message(
-    id: json['id'] as String,
-    text: json['text'] as String,
-    type: json['type'] as String,
-    attachments: (json['attachments'] as List)
-        ?.map((e) => e == null
-            ? null
-            : Attachment.fromJson((e as Map)?.map(
-                (k, e) => MapEntry(k as String, e),
-              )))
-        ?.toList(),
-    mentionedUsers: (json['mentioned_users'] as List)
-        ?.map((e) => e == null
-            ? null
-            : User.fromJson((e as Map)?.map(
-                (k, e) => MapEntry(k as String, e),
-              )))
-        ?.toList(),
-    silent: json['silent'] as bool,
-    shadowed: json['shadowed'] as bool,
-    reactionCounts: (json['reaction_counts'] as Map)?.map(
-      (k, e) => MapEntry(k as String, e as int),
+    id: json['id'] as String?,
+    text: json['text'] as String?,
+    type: json['type'] as String? ?? 'regular',
+    attachments: (json['attachments'] as List<dynamic>?)
+            ?.map((e) => Attachment.fromJson(e as Map<String, dynamic>))
+            .toList() ??
+        [],
+    mentionedUsers: (json['mentioned_users'] as List<dynamic>?)
+            ?.map((e) => User.fromJson(e as Map<String, dynamic>))
+            .toList() ??
+        [],
+    silent: json['silent'] as bool? ?? false,
+    shadowed: json['shadowed'] as bool? ?? false,
+    reactionCounts: (json['reaction_counts'] as Map<String, dynamic>?)?.map(
+      (k, e) => MapEntry(k, e as int),
     ),
-    reactionScores: (json['reaction_scores'] as Map)?.map(
-      (k, e) => MapEntry(k as String, e as int),
+    reactionScores: (json['reaction_scores'] as Map<String, dynamic>?)?.map(
+      (k, e) => MapEntry(k, e as int),
     ),
-    latestReactions: (json['latest_reactions'] as List)
-        ?.map((e) => e == null
-            ? null
-            : Reaction.fromJson((e as Map)?.map(
-                (k, e) => MapEntry(k as String, e),
-              )))
-        ?.toList(),
-    ownReactions: (json['own_reactions'] as List)
-        ?.map((e) => e == null
-            ? null
-            : Reaction.fromJson((e as Map)?.map(
-                (k, e) => MapEntry(k as String, e),
-              )))
-        ?.toList(),
-    parentId: json['parent_id'] as String,
+    latestReactions: (json['latest_reactions'] as List<dynamic>?)
+        ?.map((e) => Reaction.fromJson(e as Map<String, dynamic>))
+        .toList(),
+    ownReactions: (json['own_reactions'] as List<dynamic>?)
+        ?.map((e) => Reaction.fromJson(e as Map<String, dynamic>))
+        .toList(),
+    parentId: json['parent_id'] as String?,
     quotedMessage: json['quoted_message'] == null
         ? null
-        : Message.fromJson((json['quoted_message'] as Map)?.map(
-            (k, e) => MapEntry(k as String, e),
-          )),
-    quotedMessageId: json['quoted_message_id'] as String,
-    replyCount: json['reply_count'] as int,
-    threadParticipants: (json['thread_participants'] as List)
-        ?.map((e) => e == null
-            ? null
-            : User.fromJson((e as Map)?.map(
-                (k, e) => MapEntry(k as String, e),
-              )))
-        ?.toList(),
-    showInChannel: json['show_in_channel'] as bool,
-    command: json['command'] as String,
+        : Message.fromJson(json['quoted_message'] as Map<String, dynamic>),
+    quotedMessageId: json['quoted_message_id'] as String?,
+    replyCount: json['reply_count'] as int?,
+    threadParticipants: (json['thread_participants'] as List<dynamic>?)
+        ?.map((e) => User.fromJson(e as Map<String, dynamic>))
+        .toList(),
+    showInChannel: json['show_in_channel'] as bool?,
+    command: json['command'] as String?,
     createdAt: json['created_at'] == null
         ? null
         : DateTime.parse(json['created_at'] as String),
@@ -72,10 +52,8 @@ Message _$MessageFromJson(Map json) {
         : DateTime.parse(json['updated_at'] as String),
     user: json['user'] == null
         ? null
-        : User.fromJson((json['user'] as Map)?.map(
-            (k, e) => MapEntry(k as String, e),
-          )),
-    pinned: json['pinned'] as bool,
+        : User.fromJson(json['user'] as Map<String, dynamic>),
+    pinned: json['pinned'] as bool? ?? false,
     pinnedAt: json['pinned_at'] == null
         ? null
         : DateTime.parse(json['pinned_at'] as String),
@@ -84,16 +62,12 @@ Message _$MessageFromJson(Map json) {
         : DateTime.parse(json['pin_expires'] as String),
     pinnedBy: json['pinned_by'] == null
         ? null
-        : User.fromJson((json['pinned_by'] as Map)?.map(
-            (k, e) => MapEntry(k as String, e),
-          )),
-    extraData: (json['extra_data'] as Map)?.map(
-      (k, e) => MapEntry(k as String, e),
-    ),
+        : User.fromJson(json['pinned_by'] as Map<String, dynamic>),
+    extraData: json['extra_data'] as Map<String, dynamic>? ?? {},
     deletedAt: json['deleted_at'] == null
         ? null
         : DateTime.parse(json['deleted_at'] as String),
-    skipPush: json['skip_push'] as bool,
+    skipPush: json['skip_push'] as bool? ?? false,
   );
 }
 
@@ -110,8 +84,7 @@ Map<String, dynamic> _$MessageToJson(Message instance) {
   }
 
   writeNotNull('type', readonly(instance.type));
-  writeNotNull(
-      'attachments', instance.attachments?.map((e) => e?.toJson())?.toList());
+  val['attachments'] = instance.attachments.map((e) => e.toJson()).toList();
   val['mentioned_users'] = Serialization.userIds(instance.mentionedUsers);
   writeNotNull('reaction_counts', readonly(instance.reactionCounts));
   writeNotNull('reaction_scores', readonly(instance.reactionScores));
@@ -134,15 +107,15 @@ Map<String, dynamic> _$MessageToJson(Message instance) {
   val['pinned_at'] = readonly(instance.pinnedAt);
   val['pin_expires'] = instance.pinExpires?.toIso8601String();
   val['pinned_by'] = readonly(instance.pinnedBy);
-  writeNotNull('extra_data', instance.extraData);
+  val['extra_data'] = instance.extraData;
   writeNotNull('deleted_at', readonly(instance.deletedAt));
   return val;
 }
 
-TranslatedMessage _$TranslatedMessageFromJson(Map json) {
+TranslatedMessage _$TranslatedMessageFromJson(Map<String, dynamic> json) {
   return TranslatedMessage(
-    (json['i18n'] as Map)?.map(
-      (k, e) => MapEntry(k as String, e as String),
+    (json['i18n'] as Map<String, dynamic>?)?.map(
+      (k, e) => MapEntry(k, e as String),
     ),
   );
 }
