@@ -6,33 +6,24 @@ part of 'own_user.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-OwnUser _$OwnUserFromJson(Map json) {
+OwnUser _$OwnUserFromJson(Map<String, dynamic> json) {
   return OwnUser(
-    devices: (json['devices'] as List)
-        ?.map((e) => e == null
-            ? null
-            : Device.fromJson((e as Map)?.map(
-                (k, e) => MapEntry(k as String, e),
-              )))
-        ?.toList(),
-    mutes: (json['mutes'] as List)
-        ?.map((e) => e == null
-            ? null
-            : Mute.fromJson((e as Map)?.map(
-                (k, e) => MapEntry(k as String, e),
-              )))
-        ?.toList(),
-    totalUnreadCount: json['total_unread_count'] as int,
-    unreadChannels: json['unread_channels'] as int,
-    channelMutes: (json['channel_mutes'] as List)
-        ?.map((e) => e == null
-            ? null
-            : Mute.fromJson((e as Map)?.map(
-                (k, e) => MapEntry(k as String, e),
-              )))
-        ?.toList(),
+    devices: (json['devices'] as List<dynamic>?)
+            ?.map((e) => Device.fromJson(e as Map<String, dynamic>))
+            .toList() ??
+        [],
+    mutes: (json['mutes'] as List<dynamic>?)
+            ?.map((e) => Mute.fromJson(e as Map<String, dynamic>))
+            .toList() ??
+        [],
+    totalUnreadCount: json['total_unread_count'] as int? ?? 0,
+    unreadChannels: json['unread_channels'] as int?,
+    channelMutes: (json['channel_mutes'] as List<dynamic>?)
+            ?.map((e) => Mute.fromJson(e as Map<String, dynamic>))
+            .toList() ??
+        [],
     id: json['id'] as String,
-    role: json['role'] as String,
+    role: json['role'] as String? ?? '',
     createdAt: json['created_at'] == null
         ? null
         : DateTime.parse(json['created_at'] as String),
@@ -42,11 +33,9 @@ OwnUser _$OwnUserFromJson(Map json) {
     lastActive: json['last_active'] == null
         ? null
         : DateTime.parse(json['last_active'] as String),
-    online: json['online'] as bool,
-    extraData: (json['extra_data'] as Map)?.map(
-      (k, e) => MapEntry(k as String, e),
-    ),
-    banned: json['banned'] as bool,
+    online: json['online'] as bool? ?? false,
+    extraData: json['extra_data'] as Map<String, dynamic>,
+    banned: json['banned'] as bool? ?? false,
   );
 }
 
@@ -67,7 +56,7 @@ Map<String, dynamic> _$OwnUserToJson(OwnUser instance) {
   writeNotNull('last_active', readonly(instance.lastActive));
   writeNotNull('online', readonly(instance.online));
   writeNotNull('banned', readonly(instance.banned));
-  writeNotNull('extra_data', instance.extraData);
+  val['extra_data'] = instance.extraData;
   writeNotNull('devices', readonly(instance.devices));
   writeNotNull('mutes', readonly(instance.mutes));
   writeNotNull('channel_mutes', readonly(instance.channelMutes));
