@@ -10,16 +10,14 @@ class Serialization {
   static const Function readOnly = readonly;
 
   /// List of users to list of userIds
-  static List<String> userIds(List<User> users) =>
-      users?.map((u) => u.id)?.toList();
+  static List<String>? userIds(List<User>? users) =>
+      users?.map((u) => u.id).toList();
 
   /// Takes unknown json keys and puts them in the `extra_data` key
   static Map<String, dynamic> moveToExtraDataFromRoot(
     Map<String, dynamic> json,
     List<String> topLevelFields,
   ) {
-    if (json == null) return null;
-
     final jsonClone = Map<String, dynamic>.from(json);
 
     final extraDataMap = Map<String, dynamic>.from(json)
@@ -38,7 +36,6 @@ class Serialization {
   /// the json map
   static Map<String, dynamic> moveFromExtraDataToRoot(
     Map<String, dynamic> json,
-    List<String> topLevelFields,
   ) {
     final jsonClone = Map<String, dynamic>.from(json);
     return jsonClone
