@@ -1,5 +1,5 @@
-import 'package:test/test.dart';
 import 'package:stream_chat/stream_chat.dart';
+import 'package:test/test.dart';
 
 void main() {
   group('src/api/requests', () {
@@ -12,7 +12,8 @@ void main() {
     test('PaginationParams', () {
       const option = PaginationParams();
       final j = option.toJson();
-      expect(j, {'limit': 10, 'offset': 0});
+      expect(j, containsPair('limit', 10));
+      expect(j, containsPair('offset', 0));
     });
   });
 }
