@@ -1,16 +1,17 @@
 import 'dart:convert';
 
-import 'package:test/test.dart';
 import 'package:stream_chat/src/models/channel_config.dart';
 import 'package:stream_chat/src/models/channel_state.dart';
 import 'package:stream_chat/src/models/command.dart';
 import 'package:stream_chat/src/models/message.dart';
 import 'package:stream_chat/src/models/user.dart';
 import 'package:stream_chat/stream_chat.dart';
+import 'package:test/test.dart';
 
 void main() {
   group('src/models/channel_state', () {
-    const jsonExample = '''{
+    const jsonExample = '''
+  {
    "channel": {
      "id": "dev",
      "type": "team",
@@ -844,26 +845,26 @@ void main() {
 
     test('should parse json correctly', () {
       final channelState = ChannelState.fromJson(json.decode(jsonExample));
-      expect(channelState.channel.cid, 'team:dev');
-      expect(channelState.channel.id, 'dev');
-      expect(channelState.channel.team, 'test');
-      expect(channelState.channel.type, 'team');
-      expect(channelState.channel.config, isA<ChannelConfig>());
-      expect(channelState.channel.config, isNotNull);
-      expect(channelState.channel.config.commands, hasLength(1));
-      expect(channelState.channel.config.commands[0], isA<Command>());
-      expect(channelState.channel.lastMessageAt,
+      expect(channelState.channel?.cid, 'team:dev');
+      expect(channelState.channel?.id, 'dev');
+      expect(channelState.channel?.team, 'test');
+      expect(channelState.channel?.type, 'team');
+      expect(channelState.channel?.config, isA<ChannelConfig>());
+      expect(channelState.channel?.config, isNotNull);
+      expect(channelState.channel?.config.commands, hasLength(1));
+      expect(channelState.channel?.config.commands[0], isA<Command>());
+      expect(channelState.channel?.lastMessageAt,
           DateTime.parse('2020-01-30T13:43:41.062362Z'));
-      expect(channelState.channel.createdAt,
+      expect(channelState.channel?.createdAt,
           DateTime.parse('2019-04-03T18:43:33.213373Z'));
-      expect(channelState.channel.updatedAt,
+      expect(channelState.channel?.updatedAt,
           DateTime.parse('2019-04-03T18:43:33.213374Z'));
-      expect(channelState.channel.createdBy, isA<User>());
-      expect(channelState.channel.frozen, true);
-      expect(channelState.channel.extraData['example'], 1);
-      expect(channelState.channel.extraData['name'], '#dev');
+      expect(channelState.channel?.createdBy, isA<User>());
+      expect(channelState.channel?.frozen, true);
+      expect(channelState.channel?.extraData!['example'], 1);
+      expect(channelState.channel?.extraData!['name'], '#dev');
       expect(
-        channelState.channel.extraData['image'],
+        channelState.channel?.extraData!['image'],
         'https://cdn.chrisshort.net/testing-certificate-chains-in-go/GOPHER_MIC_DROP.png',
       );
       expect(channelState.messages, hasLength(25));
@@ -888,8 +889,8 @@ void main() {
               "image": "https://cdn.chrisshort.net/testing-certificate-chains-in-go/GOPHER_MIC_DROP.png",
               "example": 1
             },
-            "watchers": null,
-            "read": null,
+            "watchers": [],
+            "read": [],
             "messages": [
               {
                 "id": "dry-meadow-0-2b73cc8b-cd86-4a01-8d40-bd82ad07a030",
@@ -901,7 +902,7 @@ void main() {
                 "show_in_channel": null,
                 "mentioned_users": [],
                 "status": "SENT",
-            "skip_push": null,
+                "skip_push": false,
                 "silent": false,
                 "pinned": false,
                 "pinned_at": null,
@@ -918,7 +919,7 @@ void main() {
                 "show_in_channel": null,
                 "mentioned_users": [],
                 "status": "SENT",
-            "skip_push": null,
+                "skip_push": false,
                 "silent": false,
                 "pinned": false,
                 "pinned_at": null,
@@ -928,7 +929,7 @@ void main() {
               {
                 "id": "dry-meadow-0-53e6299f-9b97-4a9c-a27e-7e2dde49b7e0",
                 "text": "test message",
-            "skip_push": null,
+                "skip_push": false,
                 "attachments": [],
                 "parent_id": null,
                 "quoted_message": null,
@@ -951,7 +952,7 @@ void main() {
                 "quoted_message_id": null,
                 "show_in_channel": null,
                 "mentioned_users": [],
-            "skip_push": null,
+                "skip_push": false,
                 "status": "SENT",
                 "silent": false,
                 "pinned": false,
@@ -963,7 +964,7 @@ void main() {
                 "id": "dry-meadow-0-64d7970f-ede8-4b31-9738-1bc1756d2bfe",
                 "text": "test",
                 "attachments": [],
-            "skip_push": null,
+                "skip_push": false,
                 "parent_id": null,
                 "quoted_message": null,
                 "quoted_message_id": null,
@@ -981,7 +982,7 @@ void main() {
                 "text": "hi",
                 "attachments": [],
                 "parent_id": null,
-            "skip_push": null,
+                "skip_push": false,
                 "quoted_message": null,
                 "quoted_message_id": null,
                 "show_in_channel": null,
@@ -999,7 +1000,7 @@ void main() {
                 "attachments": [],
                 "parent_id": null,
                 "quoted_message": null,
-            "skip_push": null,
+                "skip_push": false,
                 "quoted_message_id": null,
                 "show_in_channel": null,
                 "mentioned_users": [],
@@ -1022,7 +1023,7 @@ void main() {
                 "status": "SENT",
                 "silent": false,
                 "pinned": false,
-            "skip_push": null,
+                "skip_push": false,
                 "pinned_at": null,
                 "pin_expires": null,
                 "pinned_by": null
@@ -1040,7 +1041,7 @@ void main() {
                 "silent": false,
                 "pinned": false,
                 "pinned_at": null,
-            "skip_push": null,
+                "skip_push": false,
                 "pin_expires": null,
                 "pinned_by": null
               },
@@ -1052,7 +1053,7 @@ void main() {
                 "quoted_message": null,
                 "quoted_message_id": null,
                 "show_in_channel": null,
-            "skip_push": null,
+                "skip_push": false,
                 "mentioned_users": [],
                 "status": "SENT",
                 "silent": false,
@@ -1070,7 +1071,7 @@ void main() {
                 "quoted_message_id": null,
                 "show_in_channel": null,
                 "mentioned_users": [],
-            "skip_push": null,
+                "skip_push": false,
                 "status": "SENT",
                 "silent": false,
                 "pinned": false,
@@ -1089,7 +1090,7 @@ void main() {
                 "mentioned_users": [],
                 "status": "SENT",
                 "silent": false,
-            "skip_push": null,
+                "skip_push": false,
                 "pinned": false,
                 "pinned_at": null,
                 "pin_expires": null,
@@ -1099,7 +1100,7 @@ void main() {
                 "id": "icy-recipe-7-935c396e-ddf8-4a9a-951c-0a12fa5bf055",
                 "text": "what are you doing?",
                 "attachments": [],
-            "skip_push": null,
+                "skip_push": false,
                 "parent_id": null,
                 "quoted_message": null,
                 "quoted_message_id": null,
@@ -1117,7 +1118,7 @@ void main() {
                 "text": "👍",
                 "attachments": [],
                 "parent_id": null,
-            "skip_push": null,
+                "skip_push": false,
                 "quoted_message": null,
                 "quoted_message_id": null,
                 "show_in_channel": null,
@@ -1133,7 +1134,7 @@ void main() {
                 "id": "snowy-credit-3-3e0c1a0d-d22f-42ee-b2a1-f9f49477bf21",
                 "text": "sdasas",
                 "attachments": [],
-            "skip_push": null,
+                "skip_push": false,
                 "parent_id": null,
                 "quoted_message": null,
                 "quoted_message_id": null,
@@ -1154,7 +1155,7 @@ void main() {
                 "quoted_message": null,
                 "quoted_message_id": null,
                 "show_in_channel": null,
-            "skip_push": null,
+                "skip_push": false,
                 "mentioned_users": [],
                 "status": "SENT",
                 "silent": false,
@@ -1167,7 +1168,7 @@ void main() {
                 "id": "snowy-credit-3-cfaf0b46-1daa-49c5-947c-b16d6697487d",
                 "text": "nhisagdhsadz",
                 "attachments": [],
-            "skip_push": null,
+                "skip_push": false,
                 "parent_id": null,
                 "quoted_message": null,
                 "quoted_message_id": null,
@@ -1186,7 +1187,7 @@ void main() {
                 "attachments": [],
                 "parent_id": null,
                 "quoted_message": null,
-            "skip_push": null,
+                "skip_push": false,
                 "quoted_message_id": null,
                 "show_in_channel": null,
                 "mentioned_users": [],
@@ -1203,7 +1204,7 @@ void main() {
                 "attachments": [],
                 "parent_id": null,
                 "quoted_message": null,
-            "skip_push": null,
+                "skip_push": false,
                 "quoted_message_id": null,
                 "show_in_channel": null,
                 "mentioned_users": [],
@@ -1211,7 +1212,7 @@ void main() {
                 "silent": false,
                 "pinned": false,
                 "pinned_at": null,
-            "skip_push": null,
+                "skip_push": false,
                 "pin_expires": null,
                 "pinned_by": null
               },
@@ -1228,7 +1229,7 @@ void main() {
                 "silent": false,
                 "pinned": false,
                 "pinned_at": null,
-            "skip_push": null,
+                "skip_push": false,
                 "pin_expires": null,
                 "pinned_by": null
               },
@@ -1245,7 +1246,7 @@ void main() {
                 "silent": false,
                 "pinned": false,
                 "pinned_at": null,
-            "skip_push": null,
+                "skip_push": false,
                 "pin_expires": null,
                 "pinned_by": null
               },
@@ -1262,7 +1263,7 @@ void main() {
                 "silent": false,
                 "pinned": false,
                 "pinned_at": null,
-            "skip_push": null,
+                "skip_push": false,
                 "pin_expires": null,
                 "pinned_by": null
               },
@@ -1279,7 +1280,7 @@ void main() {
                 "silent": false,
                 "pinned": false,
                 "pinned_at": null,
-            "skip_push": null,
+                "skip_push": false,
                 "pin_expires": null,
                 "pinned_by": null
               },
@@ -1296,7 +1297,7 @@ void main() {
                 "silent": false,
                 "pinned": false,
                 "pinned_at": null,
-            "skip_push": null,
+                "skip_push": false,
                 "pin_expires": null,
                 "pinned_by": null
               },
@@ -1313,7 +1314,7 @@ void main() {
                 "silent": false,
                 "pinned": false,
                 "pinned_at": null,
-            "skip_push": null,
+                "skip_push": false,
                 "pin_expires": null,
                 "pinned_by": null
               }
@@ -1329,10 +1330,10 @@ void main() {
         members: [],
         messages:
             (j['messages'] as List).map((m) => Message.fromJson(m)).toList(),
-        read: null,
+        read: [],
         watcherCount: 5,
         pinnedMessages: [],
-        watchers: null,
+        watchers: [],
       );
 
       expect(
