@@ -1,8 +1,8 @@
 import 'dart:convert';
 
-import 'package:test/test.dart';
 import 'package:stream_chat/src/models/reaction.dart';
 import 'package:stream_chat/src/models/user.dart';
+import 'package:test/test.dart';
 
 void main() {
   group('src/models/reaction', () {
@@ -33,8 +33,8 @@ void main() {
       expect(reaction.createdAt, DateTime.parse('2020-01-28T22:17:31.108742Z'));
       expect(reaction.type, 'wow');
       expect(
-        reaction.user.toJson(),
-        User.init('2de0297c-f3f2-489d-b930-ef77342edccf', extraData: {
+        reaction.user?.toJson(),
+        User(id: '2de0297c-f3f2-489d-b930-ef77342edccf', extraData: {
           'image': 'https://randomuser.me/api/portraits/women/45.jpg',
           'name': 'Daisy Morgan'
         }).toJson(),
@@ -49,7 +49,7 @@ void main() {
         messageId: '76cd8c82-b557-4e48-9d12-87995d3a0e04',
         createdAt: DateTime.parse('2020-01-28T22:17:31.108742Z'),
         type: 'wow',
-        user: User.init('2de0297c-f3f2-489d-b930-ef77342edccf', extraData: {
+        user: User(id: '2de0297c-f3f2-489d-b930-ef77342edccf', extraData: {
           'image': 'https://randomuser.me/api/portraits/women/45.jpg',
           'name': 'Daisy Morgan'
         }),
